Added three new wrapper functions TravelDistCreate, TurnCreate, BeepCreate
The functions travelDist, turnAngle and beep expect the robot port; this caused confusion in the lab since students work with the robot structure (that is what is created in the initialization function). To avoid this, in 2023, we added 3 wrapper functions (TravelDistCreate, TurnCreate, BeepCreate) that the student will use. The input for these functions is the robot structure. For the simulator is does not matter, but since we are adding these new functions, we need to also add them to the simulator